polyline generator used by line_bucket and DrawableBuilder
