+++ bazel/core.bzl
@@ -612,6 +612,7 @@ MLN_CORE_SOURCE = [
     "src/mbgl/util/premultiply.cpp",
     "src/mbgl/util/quaternion.cpp",
     "src/mbgl/util/quaternion.hpp",
+    "src/mbgl/util/polyline_generator.cpp",
     "src/mbgl/util/rapidjson.cpp",
     "src/mbgl/util/rapidjson.hpp",
     "src/mbgl/util/rect.hpp",
@@ -818,6 +819,7 @@ MLN_CORE_HEADERS = [
     "include/mbgl/util/monotonic_timer.hpp",
     "include/mbgl/util/noncopyable.hpp",
     "include/mbgl/util/platform.hpp",
+    "include/mbgl/util/polyline_generator.hpp",
     "include/mbgl/util/premultiply.hpp",
     "include/mbgl/util/projection.hpp",
     "include/mbgl/util/range.hpp",
