+++ include/mbgl/gfx/drawable_builder.hpp
@@ -5,6 +5,7 @@
 #include <mbgl/util/string_indexer.hpp>
 #include <mbgl/tile/geometry_tile_data.hpp>
 #include <mbgl/style/types.hpp>
+#include <mbgl/util/polyline_generator.hpp>
 
 #include <array>
 #include <memory>
@@ -38,16 +39,6 @@ public:
         Primitives, ///< building primitive drawables. Not implemented
         Polylines,  ///< building drawables for thick polylines
         Custom      ///< building custom drawables.
-    };
-
-    struct PolylineOptions {
-        FeatureType type{FeatureType::LineString};
-        style::LineJoinType joinType{style::LineJoinType::Miter};
-        float miterLimit{2.f};
-        style::LineCapType beginCap{style::LineCapType::Butt};
-        style::LineCapType endCap{style::LineCapType::Butt};
-        float roundLimit{1.f};
-        uint32_t overscaling{1}; // TODO: what is this???
     };
 
 public:
@@ -215,7 +206,7 @@ public:
     virtual std::unique_ptr<Drawable::DrawSegment> createSegment(gfx::DrawMode, SegmentBase&&) = 0;
 
     /// Add a polyline. If the last point equals the first it will be closed, otherwise open
-    void addPolyline(const GeometryCoordinates& coordinates, const PolylineOptions&);
+    void addPolyline(const GeometryCoordinates& coordinates, const util::PolylineGeneratorOptions&);
 
 protected:
     std::size_t curVertexCount() const;
