+++ src/mbgl/renderer/buckets/line_bucket.hpp
@@ -62,37 +62,6 @@ public:
 private:
     void addGeometry(const GeometryCoordinates&, const GeometryTileFeature&, const CanonicalTileID&);
 
-    struct TriangleElement {
-        TriangleElement(uint16_t a_, uint16_t b_, uint16_t c_)
-            : a(a_),
-              b(b_),
-              c(c_) {}
-        uint16_t a, b, c;
-    };
-
-    class Distances;
-    void addCurrentVertex(const GeometryCoordinate& currentCoordinate,
-                          double& distance,
-                          const Point<double>& normal,
-                          double endLeft,
-                          double endRight,
-                          bool round,
-                          std::size_t startVertex,
-                          std::vector<LineBucket::TriangleElement>& triangleStore,
-                          std::optional<Distances> distances);
-
-    void addPieSliceVertex(const GeometryCoordinate& currentVertex,
-                           double distance,
-                           const Point<double>& extrude,
-                           bool lineTurnsLeft,
-                           std::size_t startVertex,
-                           std::vector<TriangleElement>& triangleStore,
-                           std::optional<Distances> distances);
-
-    std::ptrdiff_t e1;
-    std::ptrdiff_t e2;
-    std::ptrdiff_t e3;
-
     const float zoom;
     const uint32_t overscaling;
 };
